requirement.txt
Flask
flask_cors  # need when gets cors erros
dnspython
PyJWT
pillow

